converts app to es6modules
