+++ client/zone/js-modules/vacuum-map.js
@@ -1,145 +1,9 @@
-function MapDrawer() {
-    const mapCanvas = document.createElement('canvas');
-    const mapCtx = mapCanvas.getContext("2d");
-    let boundingBox = {
-        minX: 0,
-        minY: 0,
-        maxX: 1024,
-        maxY: 1024
-    }
+import { MapDrawer } from "./map-drawer.js";
+import { PathDrawer } from "./path-drawer.js";
+import { trackTransforms } from "./tracked-canvas.js";
+import { transformFromMeter, flipX, noTransform } from "./coordinate-transforms.js";
 
-    mapCanvas.width = 1024;
-    mapCanvas.height = 1024;
-
-
-    function draw(mapData) {
-        this.boundingBox = getBoundingBox(mapData, mapCanvas.width, mapCanvas.height);
-
-        mapCtx.clearRect(0, 0, mapCanvas.width, mapCanvas.height);
-        const imgData = mapCtx.createImageData(mapCanvas.width, mapCanvas.height);
-        mapData.forEach(function (px) {
-            if (px[1] === 0 && px[2] === 0 && px[3] === 0) {
-                imgData.data[px[0]] = 102;
-                imgData.data[px[0] + 1] = 153;
-                imgData.data[px[0] + 2] = 255;
-                imgData.data[px[0] + 3] = 255;
-
-            } else if (px[1] === 255 && px[2] === 255 && px[3] === 255) {
-                imgData.data[px[0]] = 0;
-                imgData.data[px[0] + 1] = 118;
-                imgData.data[px[0] + 2] = 255;
-                imgData.data[px[0] + 3] = 255;
-
-            } else {
-                imgData.data[px[0]] = 0;
-                imgData.data[px[0] + 1] = 118;
-                imgData.data[px[0] + 2] = 255;
-                imgData.data[px[0] + 3] = 255;
-            }
-        });
-        mapCtx.putImageData(imgData, 0, 0);
-    }
-
-    function getBoundingBox(mapData, maxWidth, maxHeight) {
-        let minX = maxWidth;
-        let maxX = 0;
-        let minY = maxHeight;
-        let maxY = 0;
-
-        mapData.forEach(function (px) {
-            const pxOffset = px[0] / 4;
-            const x = pxOffset % maxWidth;
-            const y = Math.floor(pxOffset / maxWidth);
-            if (x < minX) minX = x;
-            if (x > maxX) maxX = x;
-            if (y < minY) minY = y;
-            if (y > maxY) maxY = y;
-        });
-
-        return { minX, minY, maxX, maxY };
-    }
-
-    return {
-        draw: draw,
-        canvas: mapCanvas,
-        boundingBox: boundingBox
-    };
-}
-
-function PathDrawer() {
-    let path = [];
-    const canvas = document.createElement('canvas');
-    canvas.width = 1024;
-    canvas.height = 1024;
-    // Used to draw smoother path when zoomed into the map
-    let scaleFactor = 1;
-    const maxScaleFactor = 6;
-
-    const transformFromMeter = new DOMMatrix([1, 0, 0, 1, 0, 0])
-        .translateSelf(512, 512)
-        .scaleSelf(20);
-
-    // Used to flip the path when map is flipped
-    let accountForFlip = new DOMMatrix([1, 0, 0, 1, 0, 0]);
-
-    function setFlipped(isFlipped) {
-        if (isFlipped) {
-            accountForFlip = new DOMMatrix([1, 0, 0, -1, 0, 0]);
-        } else {
-            accountForFlip = new DOMMatrix([1, 0, 0, 1, 0, 0]);
-        }
-    }
-
-    function setPath(newPath) {
-        path = newPath;
-    }
-
-    function scale(factor) {
-        const newScaleFactor = Math.min(factor, maxScaleFactor);
-        if (newScaleFactor === scaleFactor) return;
-
-        const ctx = canvas.getContext("2d");
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-
-        scaleFactor = newScaleFactor;
-        canvas.width = canvas.height = scaleFactor * 1024;
-        draw();
-    }
-
-    function draw() {
-        const ctx = canvas.getContext("2d");
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-
-        const pathTransform = new DOMMatrix().scale(scaleFactor).multiply(transformFromMeter).multiplySelf(accountForFlip);
-        let first = true;
-        ctx.beginPath();
-        ctx.lineWidth = 1;
-        ctx.strokeStyle = "#ffffff";
-        for (const coord of path) {
-            const [xMeter, yMeter] = coord;
-            const { x, y } = new DOMPoint(xMeter, yMeter).matrixTransform(pathTransform);
-            if (first) {
-                ctx.moveTo(x, y);
-                first = false;
-            }
-            else {
-                ctx.lineTo(x, y);
-            }
-        }
-        ctx.stroke();
-    }
-
-    return {
-        setPath: setPath,
-        setFlipped, setFlipped,
-        scale: scale,
-        getScaleFactor: function () { return scaleFactor; },
-        canvas: canvas,
-        draw: draw
-    }
-}
-
-function VacuumMap(canvasElement) {
+export function VacuumMap(canvasElement) {
     const canvas = canvasElement;
 
     const mapDrawer = new MapDrawer();
@@ -153,16 +17,13 @@ function VacuumMap(canvasElement) {
 
     let redrawCanvas = null;
 
-    let accountForFlip = new DOMMatrix([1, 0, 0, -1, 0, 0]);
-    const transformFromMeter = new DOMMatrix([1, 0, 0, 1, 0, 0])
-        .translateSelf(512, 512)
-        .scaleSelf(20);
+    let accountForFlip = flipX;
 
     function updateMap(data) {
         if (data.yFlipped) {
-            accountForFlip = new DOMMatrix([1, 0, 0, -1, 0, 0]);
+            accountForFlip = flipX;
         } else {
-            accountForFlip = new DOMMatrix([1, 0, 0, 1, 0, 0]);
+            accountForFlip = noTransform;
         }
         mapDrawer.draw(data.map);
         pathDrawer.setPath(data.path);
@@ -172,9 +33,9 @@ function VacuumMap(canvasElement) {
     }
 
     function convertToRealCoords(tappedPoint) {
-        mapCoordsToMeters = transformFromMeter.multiply(accountForFlip).inverse();
-        Point = tappedPoint.matrixTransform(mapCoordsToMeters);
-        [x1Real, y1Real] = [Point.x, Point.y].map(x => Math.round(-1000 * x));
+        const mapCoordsToMeters = transformFromMeter.multiply(accountForFlip).inverse();
+        const point = tappedPoint.matrixTransform(mapCoordsToMeters);
+        const [x1Real, y1Real] = [point.x, point.y].map(x => Math.round(-1000 * x));
         return { 'x': x1Real, 'y': y1Real };
     }
 
@@ -414,74 +275,6 @@ function VacuumMap(canvasElement) {
         canvas.addEventListener('DOMMouseScroll', handleScroll, false);
         canvas.addEventListener('mousewheel', handleScroll, false);
     };
-
-    function trackTransforms(ctx) {
-        let xform = new DOMMatrix();
-        ctx.getTransform = function () { return xform; };
-
-        const savedTransforms = [];
-        const save = ctx.save;
-        ctx.save = function () {
-            savedTransforms.push(xform.translate(0, 0));
-            return save.call(ctx);
-        };
-
-        const restore = ctx.restore;
-        ctx.restore = function () {
-            xform = savedTransforms.pop();
-            return restore.call(ctx);
-        };
-
-        const scale = ctx.scale;
-        ctx.scale = function (sx, sy) {
-            xform = xform.scale(sx, sy);
-            return scale.call(ctx, sx, sy);
-        };
-
-        const rotate = ctx.rotate;
-        ctx.rotate = function (radians) {
-            xform = xform.rotate(radians * 180 / Math.PI);
-            return rotate.call(ctx, radians);
-        };
-
-        const translate = ctx.translate;
-        ctx.translate = function (dx, dy) {
-            xform = xform.translate(dx, dy);
-            return translate.call(ctx, dx, dy);
-        };
-
-        const transform = ctx.transform;
-        ctx.transform = function (a, b, c, d, e, f) {
-            const m2 = new DOMMatrix();
-            m2.a = a; m2.b = b; m2.c = c; m2.d = d; m2.e = e; m2.f = f;
-            xform = xform.multiply(m2);
-            return transform.call(ctx, a, b, c, d, e, f);
-        };
-
-        const setTransform = ctx.setTransform;
-        ctx.setTransform = function (a, b, c, d, e, f) {
-            xform.a = a;
-            xform.b = b;
-            xform.c = c;
-            xform.d = d;
-            xform.e = e;
-            xform.f = f;
-            return setTransform.call(ctx, a, b, c, d, e, f);
-        };
-
-        const pt = new DOMPoint();
-        ctx.transformedPoint = function (x, y) {
-            pt.x = x; pt.y = y;
-            return pt.matrixTransform(xform.inverse());
-        }
-
-        ctx.getScaleFactor2d = function () {
-            const sx = Math.sqrt(xform.a * xform.a + xform.b + xform.b);
-            const sy = Math.sqrt(xform.c * xform.c + xform.d * xform.d);
-
-            return [sx, sy];
-        }
-    }
 
     return {
         initCanvas: initCanvas,
